requirements.txt
pytube==12.1.0
streamlit==1.9.2
requests==2.27.1
Pillow==9.0.1